fix ex00

+++ CPP09/ex00/headers/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: hgeissle <hgeissle@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 14:10:57 by hgeissle          #+#    #+#             */
-/*   Updated: 2023/09/12 11:43:57 by hgeissle         ###   ########.fr       */
+/*   Updated: 2023/09/12 15:29:51 by hgeissle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ class BitcoinExchange
 		BitcoinExchange &operator=(const BitcoinExchange &src);
 		~BitcoinExchange();
 
-		static std::map<std::string, float> _database;
+		static std::map<std::string, double> _database;
 
 		static bool _isDateValid(std::string date);
 		static bool _isDateFormatValid(std::string date);
